Fixes a bug that occurred when a logic function was decorated.
This fixes issue #12.  If a logic function was decorated and you sent a
request to a doctor endpoint using the content-type application/json and
that json body had a parameter not in the signature of the logic
function, then a TypeError would occur as it would attempt to pass the
extra parameter as a **kwarg to the logic function.