Don't try to parse json body on GET request.
Technically JSON bodies are allowed on GET/DELETE requests, but
according to the HTTP spec, they have no contextual meaning and
shouldn't alter the response.  Because of this the body can't be parsed
in flask.  If this happens we instead try to parse parameters from the
query string and form.