Fixes issue 18.
This fixes a bug introducded in v1.1.3.  This bug would only occur if a
logic function was decorated and that decorator passed a positional
argument to the logic function.  Doctor would think the positional
argument passed by the decorator was a required request parameter even
if it was specified to be omitted in the router using `omit_args`.