Fix more things for Python 3
It turns out Python 3 tests weren't running in Travis, and a local
misconfiguration issue was causing the Python 3 tests to be run
using Python 2 instead, so not all the Python 3 things were fixed
in 1.2.0.

One significant change in this release is that unicode strings
aren't encoded as UTF-8 anymore. I think that makes the most sense
for Python 3, and this way both Python 2 and 3 work the same way.
But if we want to keep strict compatibility with the way things
were working in Python 2 before, we could encode to UTF-8 for only
Python 2 instead...