Updated the JsonSchema type to work properly.
Previously this type did not work with doctor since it required
instantiating an instance in order to obtain the example and description
values of the type.  Now the `json_schema_type` helper function will
load the schema and set those attributes on the generated class for you.

I didn't update the example app, but locally I tested out all the json
schema types and they properly get converted to the right python types.