Removing the code for documenting object properties.
The reference link now handles this.  This ends up cleaning up the
actual request/response documentation.  The documenting of object
properties was also problematic if an object was in a query parameter or
if a property was also an object.

This also means we can remove our dependency on our forked version of
httpdomain.py