Fixed issue discovered while testing flask==1.0.2
It looks like in MethodView the methods attribute was always a set if
methods was not provided.  Prior to v1.0.0 we were passing it in as a
list and the MethodView kept it that way instead of converting to a set.
With >= v1.0.0 methods will always be a set.  To make this work with
older versions and new versions, we now pass in methods initially as a
set and update the routing code to use .add instead of .append for the
methods of the handler.